reorganize the CatGen utility functions into one directory
