+++ catkit/gen/utils/geometry.py
@@ -1,6 +1,5 @@
+from .. import defaults
 import numpy as np
-from numpy.linalg import norm
-from . import defaults
 
 
 def matching_sites(position, comparators, tol=1e-8):
@@ -34,36 +33,6 @@ def matching_sites(position, comparators, tol=1e-8):
     return match
 
 
-def matching_coordinates(position, comparators, tol=1e-8):
-    """Get the indices of all points in a comparator list that are
-    equal to a given position (with a tolerance), taking into
-    account periodic boundary conditions (adaptation from Pymatgen).
-
-    This will only accept a Cartesian coordinate scheme.
-    TODO: merge this with matching_sites.
-
-    Parameters
-    ----------
-    position : list (3,)
-        Fractional coordinate to compare to list.
-    comparators : list (3, n)
-        Fractional coordinates to compare against.
-    tol : float
-        Absolute tolerance.
-
-    Returns
-    -------
-    match : list (n,)
-        Indices of matches.
-    """
-    if len(comparators) == 0:
-        return []
-
-    fdist = comparators - position[None, :]
-    match = np.where((np.abs(fdist) < tol).all(axis=1))[0]
-
-    return match
-
 def _get_basis_vectors(coordinates):
     if len(coordinates) == 3:
         c0, c1, c2 = coordinates
@@ -75,9 +44,9 @@ def _get_basis_vectors(coordinates):
     basis2 = np.cross(basis1, c0 - c2)
     basis3 = np.cross(basis1, basis2)
 
-    basis1 /= norm(basis1)
-    basis2 /= norm(basis2)
-    basis3 /= norm(basis3)
+    basis1 /= np.linalg.norm(basis1)
+    basis2 /= np.linalg.norm(basis2)
+    basis3 /= np.linalg.norm(basis3)
 
     basis_vectors = np.vstack([basis1, basis2, basis3])
 
@@ -96,7 +65,7 @@ def _get_position(coordinates, basis=None,
     basis[2] *= np.cos(tor)
 
     vector = basis[1] + basis[2]
-    vector /= norm(vector)
+    vector /= np.linalg.norm(vector)
 
     vector *= distance * np.sin(ang)
     basis[0] *= distance * np.cos(ang)
